+++ Sources/PFMMacros/Plugin.swift
@@ -1,0 +1,10 @@
+import SwiftCompilerPlugin
+import SwiftSyntaxMacros
+
+@main
+struct PFMMacrosPlugin: CompilerPlugin {
+    let providingMacros: [Macro.Type] = [
+        GenerableMacro.self,
+        GuideMacro.self,
+    ]
+}